backend/requirements.txt
python-dotenv
fastapi
uvicorn
nltk
rake-nltk
sumy